backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
motor==3.3.2
pymongo==4.6.0 